Add database into Data layer (Room)

+++ app/src/main/java/com/tossdesu/bankcardinfo/data/network/entity/CardCountryDto.kt
@@ -4,6 +4,6 @@ data class CardCountryDto(
     val alpha2: String,
     val name: String,
     val currency: String,
-    val latitude: Int,
-    val longitude: Int
+    val latitude: Float,
+    val longitude: Float
 )
